+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Reduce model initialization compile time [#1212](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1212)
+
 ## v0.22.1
 
 - WhichTransform for default MatrixSpectralTransform on GPU at low resolution [#1194](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1194)